fix: port QueueSource and getLocalQueueLabel to hardware-profiles-shared
Components in the shared package were stale — missing the queueSource
prop and related logic added to main. Port QueueSource enum and
getLocalQueueLabel to the shared package and update shims to re-export.
