+++ packages/hardware-profiles-shared/src/HardwareProfileDetailsPopover.tsx
@@ -23,10 +23,13 @@ import {
   formatNodeSelector,
   formatIdentifierDetails,
   sortIdentifiers,
+  getLocalQueueLabel,
 } from './utils';
+import { QueueSource } from './const';
 
 type HardwareProfileDetailsPopoverProps = {
   localQueueName?: string;
+  queueSource?: QueueSource;
   priorityClass?: string;
   tolerations?: Toleration[];
   nodeSelector?: NodeSelector;
@@ -36,6 +39,7 @@ type HardwareProfileDetailsPopoverProps = {
 
 const HardwareProfileDetailsPopover: React.FC<HardwareProfileDetailsPopoverProps> = ({
   localQueueName,
+  queueSource,
   priorityClass,
   tolerations,
   nodeSelector,
@@ -102,14 +106,16 @@ const HardwareProfileDetailsPopover: React.FC<HardwareProfileDetailsPopoverProps
                   </StackItem>
                 ))}
             </>
-          ) : (
+          ) : !localQueueName ? (
             <StackItem>
               No matching hardware profile found, using existing settings. Default, min, and max
               values are not available.
             </StackItem>
-          )}
+          ) : null}
           {localQueueName && (
-            <StackItem>{renderSection('Local queue', [localQueueName])}</StackItem>
+            <StackItem>
+              {renderSection(getLocalQueueLabel(queueSource), [localQueueName])}
+            </StackItem>
           )}
           {clusterQueueName && (
             <StackItem>{renderSection('Cluster queue', [clusterQueueName])}</StackItem>
