+++ packages/hardware-profiles-shared/src/index.ts
@@ -18,6 +18,7 @@ export {
   HARDWARE_PROFILE_BINDING_CONFIG,
   MODEL_SERVING_VISIBILITY,
   INFERENCE_SERVICE_HARDWARE_PROFILE_PATHS,
+  QueueSource,
 } from './const';
 
 // Kueue
@@ -59,6 +60,7 @@ export {
   getExistingHardwareProfileData,
   assemblePodSpecOptions,
   applyHardwareProfileConfig,
+  getLocalQueueLabel,
 } from './utils';
 
 // Components
